test(e2e): map retired P1 selectors
