+++ test/onboard-inference-reconciliation.test.ts
@@ -934,77 +934,4 @@ console.log(JSON.stringify({
       );
     });
   });
-  it("migrates only allowlisted legacy credentials through mocked gateway registration (#7617)", async () => {
-    const tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), "nemoclaw-onboard-resume-cred-"));
-    const legacyDir = path.join(tmpDir, ".nemoclaw");
-    const legacyFile = path.join(legacyDir, "credentials.json");
-    fs.mkdirSync(legacyDir, { recursive: true, mode: 0o700 });
-    fs.writeFileSync(
-      legacyFile,
-      JSON.stringify({
-        OPENAI_API_KEY: "sk-TEST-NOT-A-REAL-STORED-KEY",
-        OPENSHELL_GATEWAY: "tampered-gateway",
-        NODE_OPTIONS: "--require=/tmp/tampered.js",
-      }),
-      { mode: 0o600 },
-    );
-    const credentialEnv =
-      require("../src/lib/onboard/credential-env") as typeof import("../src/lib/onboard/credential-env.js");
-    const credentialStore =
-      require("../src/lib/credentials/store") as typeof import("../src/lib/credentials/store.js");
-    try {
-      await withProcessEnv(
-        {
-          HOME: tmpDir,
-          OPENAI_API_KEY: undefined,
-          OPENSHELL_GATEWAY: "trusted-gateway",
-          NODE_OPTIONS: "--enable-source-maps",
-        },
-        async () => {
-          const harness = createDirectSetupInferenceHarness({
-            runOpenshell: (args) =>
-              args.slice(0, 2).join(" ") === "provider get"
-                ? { status: 0, stdout: "", stderr: "" }
-                : undefined,
-            overrides: { hydrateCredentialEnv: credentialEnv.hydrateCredentialEnv },
-          });
-
-          await harness.setupInference(
-            "test-box",
-            "gpt-5.4",
-            "openai-api",
-            "https://api.openai.com/v1",
-            "OPENAI_API_KEY",
-          );
-
-          assert.equal(process.env.OPENAI_API_KEY, "sk-TEST-NOT-A-REAL-STORED-KEY");
-          assert.equal(process.env.OPENSHELL_GATEWAY, "trusted-gateway");
-          assert.equal(process.env.NODE_OPTIONS, "--enable-source-maps");
-          assert.equal(
-            fs.existsSync(legacyFile),
-            true,
-            "legacy credentials.json must survive the staging-only hydrate path",
-          );
-          const providerUpdate = harness.commands.find((entry) =>
-            entry.command.includes("provider update -g nemoclaw openai-api"),
-          );
-          assert.ok(providerUpdate, "expected provider update command");
-          assert.equal(providerUpdate.env?.OPENAI_API_KEY, "sk-TEST-NOT-A-REAL-STORED-KEY");
-          assert.doesNotMatch(providerUpdate.command, /sk-TEST-NOT-A-REAL-STORED-KEY/);
-          assert.doesNotMatch(providerUpdate.command, /tampered-gateway|tampered\.js/);
-          assert.notEqual(providerUpdate.env?.OPENSHELL_GATEWAY, "tampered-gateway");
-          assert.notEqual(providerUpdate.env?.NODE_OPTIONS, "--require=/tmp/tampered.js");
-
-          credentialStore.removeLegacyCredentialsFile();
-          assert.equal(
-            fs.existsSync(legacyFile),
-            false,
-            "legacy credentials.json is removed only after mocked gateway registration succeeds",
-          );
-        },
-      );
-    } finally {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    }
-  });
 });
